+++ src/components/App.js
@@ -1,5 +1,5 @@
 import showMain from './showMain.js';
-import MenuButtonController from './MenuButtonController.js';
+import MenuButtonController from './MainMenu/MenuButtonController.js';
 import PurchaseEvent from './PurchaseEvent.js';
 
 export default class App {
